+++ audiomate/encoding/__init__.py
@@ -1,0 +1,8 @@
+"""
+The encoding module provides functionality to encode labels to use for example for training a DNN.
+"""
+
+from .base import Encoder  # noqa: F401
+
+from .frame_based import FrameHotEncoder  # noqa: F401
+from .frame_based import FrameOrdinalEncoder  # noqa: F401